A shaft-fitting app
This is the HTML and CSS for a golf shaft fitting app I've wanted to design.  It's not 100% mine; I got some help from another source.

The banner image won't load since this was written in Notepad++ and the image file is saved there.  Still, I'm pretty happy with it since there was some debugging I had to do after some faulty layout wording!